Move tx verification tests to transaction crate
